Using JPA instead of Hardcoded Todos list -backend

+++ backend/restful-web-services/src/main/java/com/macakosoft/rest/webservices/restfulwebservices/service/TodoHarcodedService.java
@@ -11,7 +11,7 @@ import com.macakosoft.rest.webservices.restfulwebservices.bean.Todo;
 @Service
 public class TodoHarcodedService {
   private static List<Todo> todos = new ArrayList<>();
-  private static int idCounter = 0;
+  private static long idCounter = 0;
 
   static {
     todos.add(new Todo(++idCounter, "macako", "learn react js", new Date(), false));
